+++ FinalTwinkie/FinalTwinkie/GameObjects/Tank.cpp
@@ -116,19 +116,18 @@ void CTank::Render(void)
 		//Health
 		RECT rect;
 		rect.top=(LONG)360; rect.left=(LONG)11; rect.right=(LONG)177; rect.bottom=(LONG)381;
-		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (GetPosX()-41)+pCam->GetPosX(), (GetPosY()-75)+pCam->GetPosY(), 0.5f,0.5f,&rect);
+		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (int)((GetPosX()-41)+pCam->GetPosX()), (int)((GetPosY()-75)+pCam->GetPosY()), 0.5f,0.5f,&rect);
 		rect.top=(LONG)364; rect.left=(LONG)194; rect.right=(LONG)(194+((353-194)*(GetHealth()/GetMaxHealth()))); rect.bottom=(LONG)378;
-		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (GetPosX()-41)+pCam->GetPosX(), (GetPosY()-75)+pCam->GetPosY(), 0.5f,0.5f,&rect);
+		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (int)((GetPosX()-41)+pCam->GetPosX()), (int)((GetPosY()-75)+pCam->GetPosY()), 0.5f,0.5f,&rect);
 		rect.top=(LONG)391; rect.left=(LONG)351; rect.right=(LONG)(351+((509-351)*(GetHealth()/GetMaxHealth()))); rect.bottom=(LONG)403;
-		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (GetPosX()-41)+pCam->GetPosX(), (GetPosY()-75)+pCam->GetPosY(), 0.5f,0.5f,&rect);
+		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (int)((GetPosX()-41)+pCam->GetPosX()), (int)((GetPosY()-75)+pCam->GetPosY()), 0.5f,0.5f,&rect);
 	}
 
 	if(GetPosX()+pCam->GetPosX() >= -100 && GetPosX()+pCam->GetPosX() <= CGame::GetInstance()->GetWidth()+100 && GetPosY()+pCam->GetPosY() >= -100 && GetPosY()+pCam->GetPosY() <= CGame::GetInstance()->GetHeight()+100)
 	{
-	Camera* C=Camera::GetInstance();
-	CSGD_TextureManager::GetInstance()->Draw(GetImageID(), (int)(GetPosX()-GetWidth()/2+C->GetPosX()), (int)(GetPosY()-GetHeight()/2+C->GetPosY()), 1.0f, 1.0f, 0, (float)GetWidth()/2, (float)GetHeight()/2, m_fRotation);
+		Camera* C=Camera::GetInstance();
+		CSGD_TextureManager::GetInstance()->Draw(GetImageID(), (int)(GetPosX()-GetWidth()/2+C->GetPosX()), (int)(GetPosY()-GetHeight()/2+C->GetPosY()), 1.0f, 1.0f, 0, (float)GetWidth()/2, (float)GetHeight()/2, m_fRotation);
 	}
-	//CSGD_Direct3D::GetInstance()->DrawRect(GetRect(), 0,255,0);
 }
 bool CTank::CheckCollision(IEntity* pBase)
 {
